move masking operations to separate script
